long int

+++ 0x00-hello_world/6-size.c
@@ -10,7 +10,7 @@ int main(void)
 	printf("Size of a char: %c byte(s)\n", sizeof(char));
 	printf("Size of an int: %i byte(s)\n", sizeof(int));
 	printf("Size of a long int: %lu byte(s)\n", sizeof(long));
-	printf("Size of a long long int: %lli bytes(s)\n", sizeof(long long));
-	printf("Size of a float: %f bytes(s)\n", sizeof(float));
+	printf("Size of a long long int: %zu bytes(s)\n", (unsigned long)sizeof(long long));
+	printf("Size of a float: %zu bytes(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
